Put class in its own file
